+++ src/lib.rs
@@ -162,7 +162,8 @@
 	clippy::similar_names,
 	clippy::tabs_in_doc_comments,
 	clippy::len_without_is_empty,
-	clippy::needless_late_init
+	clippy::needless_late_init,
+	clippy::type_complexity
 )]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]